third_party/blink/web_tests/svg/dynamic-updates/SVGFECompositeElement-dom-operator-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'operator' attribute of the SVGFECompositeElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS overComposite1.getAttribute('operator') is "in"
PASS overComposite1.getAttribute('operator') is "arithmetic"
PASS successfullyParsed is true

TEST COMPLETE
